refactor/filter: Move filter implementations into utils.search_engine.filters
